Move common util to separate util class
