+++ openapi-cli/src/main/java/io/ballerina/openapi/generators/openapi/OpenApiConverter.java
@@ -35,19 +35,14 @@ import io.ballerina.projects.Project;
 import io.ballerina.projects.ProjectKind;
 import io.ballerina.projects.directory.ProjectLoader;
 
-import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Locale;
-import java.util.Objects;
 import java.util.Optional;
 
-import static io.ballerina.openapi.converter.Constants.JSON_EXTENSION;
-import static io.ballerina.openapi.converter.Constants.YAML_EXTENSION;
+import static io.ballerina.openapi.converter.utils.CodegenUtils.resolveContractFileName;
 
 /**
  * OpenApi related utility classes.
@@ -102,7 +97,7 @@ public class OpenApiConverter {
         syntaxTree = doc.syntaxTree();
         semanticModel = project.currentPackage().getCompilation().getSemanticModel(docId.moduleId());
         List<OASResult> openAPIDefinitions = ServiceToOpenAPIConverterUtils.generateOAS3Definition(syntaxTree,
-                semanticModel, serviceName, needJson, outPath, inputPath);
+                semanticModel, serviceName, needJson, inputPath);
 
         if (!openAPIDefinitions.isEmpty()) {
             for (OASResult definition : openAPIDefinitions) {
@@ -127,57 +122,5 @@ public class OpenApiConverter {
                 }
             }
         }
-    }
-
-    /**
-     * This method use for checking the duplicate files.
-     *
-     * @param outPath     output path for file generated
-     * @param openApiName given file name
-     * @return file name with duplicate number tag
-     */
-    private static String resolveContractFileName(Path outPath, String openApiName, Boolean isJson) {
-        if (outPath != null && Files.exists(outPath)) {
-            final File[] listFiles = new File(String.valueOf(outPath)).listFiles();
-            if (listFiles != null) {
-                openApiName = checkAvailabilityOfGivenName(openApiName, listFiles, isJson);
-            }
-        }
-        return openApiName;
-    }
-
-    private static String checkAvailabilityOfGivenName(String openApiName, File[] listFiles, Boolean isJson) {
-        for (File file : listFiles) {
-            if (System.console() != null && file.getName().equals(openApiName)) {
-                String userInput = System.console().readLine("There is already a file named ' " + file.getName() +
-                        "' in the target location. Do you want to overwrite the file? [y/N] ");
-                if (!Objects.equals(userInput.toLowerCase(Locale.ENGLISH), "y")) {
-                    openApiName = setGeneratedFileName(listFiles, openApiName, isJson);
-                }
-            }
-        }
-        return openApiName;
-    }
-
-    /**
-     * This method for setting the file name for generated file.
-     *
-     * @param listFiles      generated files
-     * @param fileName       File name
-     */
-    private static String setGeneratedFileName(File[] listFiles, String fileName, boolean isJson) {
-        int duplicateCount = 0;
-        for (File listFile : listFiles) {
-            String listFileName = listFile.getName();
-            if (listFileName.contains(".") && ((listFileName.split("\\.")).length >= 2)
-                    && (listFileName.split("\\.")[0]
-                    .equals(fileName.split("\\.")[0]))) {
-                duplicateCount++;
-            }
-        }
-        if (isJson) {
-            return fileName.split("\\.")[0] + "." + duplicateCount + JSON_EXTENSION;
-        }
-        return fileName.split("\\.")[0] + "." + duplicateCount + YAML_EXTENSION;
     }
 }
