+++ openapi-cli/src/test/resources/ballerina-to-openapi/advance/openapi/multiple_service_03.yaml
@@ -1,9 +1,9 @@
 openapi: 3.0.1
 info:
-  title: Multiple Services Without Base Path Openapi Yaml
+  title: Multiple Services No Base Path Openapi Yaml
   version: 0.0.0
 servers:
-  - url: "{server}:{port}/"
+  - url: "{server}:{port}"
     variables:
       server:
         default: petstore.swagger.io
